refactor(pipelines/cdp): modularize pipeline components and extract data simulator
- Split 673-line customer_data_platform.py into cohesive, single-responsibility modules:
  - cdp_pipeline/schemas.py: BigQuery schema loading and defaults
  - cdp_pipeline/parsing.py: ParseRecordDoFn and AssignEventTimestampDoFn
  - cdp_pipeline/sessionization.py: ProcessCustomerSessionDoFn, legacy joins
  - cdp_pipeline/sinks.py: BigQuery Storage Write API sink builders
  - cdp_pipeline/pipeline.py: build_pipeline DAG assembly and create_and_run_pipeline
  - cdp_pipeline/customer_data_platform.py: Backward-compatible facade re-exporting all symbols
- Extract data generation out of Beam worker package into standalone simulator:
  - simulator/generator.py: Synthetic shopping basket and event generation
  - simulator/publisher.py: Async Pub/Sub streaming publisher
  - scripts/03_publish_events.py: Standardized CLI launcher script
  - cdp_pipeline/generate_transaction_data.py: Backward-compatible shim forwarding to simulator
- Add unit tests for simulator and facade export parity (24/24 tests passing).
- Maintain 10.00/10 PyLint score and full compliance with Google Python style.

TAG=agy
CONV=69d55818-9356-47fa-9e0d-cd9e5b0bfe03
