+++ pipelines/cdp/cdp_pipeline/generate_transaction_data.py
@@ -11,247 +11,35 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-"""A session-aware data generator for the Customer Data Platform analytics pipeline."""
+"""Backward-compatible entry point forwarding to simulator.publisher."""
 
-import argparse
-import asyncio
-from datetime import datetime, timedelta, timezone
-import json
-import logging
 import os
-import random
-from typing import Any, Dict, List, Optional, Tuple
+import sys
+import warnings
 
-from google.cloud import pubsub_v1
+# Ensure cdp root directory is on Python path if executed directly as a script
+_CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
+_PARENT_DIR = os.path.dirname(_CURRENT_DIR)
+if _PARENT_DIR not in sys.path:
+  sys.path.insert(0, _PARENT_DIR)
 
-from cdp_pipeline.models import (
-    CouponRedemption,
-    CustomerInteractionEvent,
-    EventType,
-    TransactionItem,
+# pylint: disable=wrong-import-position
+from simulator.generator import generate_synthetic_session_events
+from simulator.publisher import get_topic_path, publish_events_to_pubsub
+
+warnings.warn(
+    "cdp_pipeline.generate_transaction_data has moved to "
+    "simulator.publisher and scripts/03_publish_events.py.",
+    DeprecationWarning,
+    stacklevel=2,
 )
 
-
-def get_topic_path(publisher: pubsub_v1.PublisherClient, project: str,
-                   topic: str) -> str:
-  """Returns a fully-qualified Pub/Sub topic path."""
-  if topic.startswith("projects/"):
-    return topic
-  return publisher.topic_path(project, topic)
-
-
-def generate_synthetic_session_events(
-    household_key: str,
-    base_tx_id: int,
-    session_offset_sec: int = 0,
-) -> Tuple[List[Dict[str, Any]], List[Dict[str, Any]]]:
-  """Generates realistic transaction items and coupons for a customer shopping session."""
-  tx_id = str(base_tx_id)
-  event_time = datetime.now(
-      timezone.utc) - timedelta(seconds=session_offset_sec)
-  now_iso = event_time.isoformat()
-  store_id = str(random.choice([101, 204, 305, 436]))
-
-  # Products in session basket
-  products = [
-      {
-          "id": "941769",
-          "price": 3.99,
-          "qty": 1,
-          "disc": 0.50
-      },
-      {
-          "id": "910635",
-          "price": 2.99,
-          "qty": 2,
-          "disc": 0.00
-      },
-      {
-          "id": "1082185",
-          "price": 1.49,
-          "qty": 1,
-          "disc": 0.25
-      },
-  ]
-  selected = random.sample(products, k=random.randint(1, len(products)))
-
-  transactions = []
-  for p in selected:
-    tx_item = TransactionItem(
-        product_id=p["id"],
-        quantity=p["qty"],
-        sales_value=round(p["price"] * p["qty"], 2),
-        store_id=store_id,
-        retail_disc=0.0,
-        coupon_disc=p["disc"],
-        coupon_match_disc=0.0,
-        day=421,
-        week_no=8,
-        trans_time="1456",
-    )
-    tx_event = CustomerInteractionEvent(
-        event_type=EventType.TRANSACTION.value,
-        household_key=household_key,
-        transaction_id=tx_id,
-        event_timestamp=now_iso,
-        transaction=tx_item,
-        coupon=None,
-    )
-    transactions.append({
-        "household_key": tx_event.household_key,
-        "transaction_id": tx_event.transaction_id,
-        "product_id": tx_item.product_id,
-        "quantity": tx_item.quantity,
-        "sales_value": tx_item.sales_value,
-        "store_id": tx_item.store_id,
-        "retail_disc": tx_item.retail_disc,
-        "coupon_disc": tx_item.coupon_disc,
-        "coupon_match_disc": tx_item.coupon_match_disc,
-        "day": tx_item.day,
-        "week_no": tx_item.week_no,
-        "trans_time": tx_item.trans_time,
-        "event_timestamp": tx_event.event_timestamp,
-    })
-
-  coupons = []
-  if random.random() < 0.7:  # 70% chance of coupon redemption
-    coupon_item = CouponRedemption(
-        coupon_upc=str(random.choice([10000085364, 51700010076, 10000089277])),
-        campaign=str(random.choice([2200, 18, 500])),
-        day=421,
-    )
-    cp_event = CustomerInteractionEvent(
-        event_type=EventType.COUPON.value,
-        household_key=household_key,
-        transaction_id=tx_id,
-        event_timestamp=now_iso,
-        transaction=None,
-        coupon=coupon_item,
-    )
-    coupons.append({
-        "household_key": cp_event.household_key,
-        "transaction_id": cp_event.transaction_id,
-        "coupon_upc": coupon_item.coupon_upc,
-        "campaign": coupon_item.campaign,
-        "day": coupon_item.day,
-        "event_timestamp": cp_event.event_timestamp,
-    })
-
-  return transactions, coupons
-
-
-async def publish_events_to_pubsub(
-    project_id: Optional[str] = None,
-    transactions_topic: Optional[str] = None,
-    coupons_topic: Optional[str] = None,
-    continuous: bool = False,
-    interval: float = 1.0,
-    count: int = 10,
-    inject_errors: bool = False,
-):
-  """Publishes sessionized transactions and coupon redemptions to Pub/Sub topics."""
-  project_id = project_id or os.environ.get("PROJECT", "<project_id>")
-  tx_topic_name = transactions_topic or os.environ.get("TRANSACTIONS_TOPIC",
-                                                       "cdp-transactions")
-  cp_topic_name = coupons_topic or os.environ.get("COUPON_REDEMPTION_TOPIC",
-                                                  "cdp-coupon-redemption")
-
-  publisher = pubsub_v1.PublisherClient()
-  tx_path = get_topic_path(publisher, project_id, tx_topic_name)
-  cp_path = get_topic_path(publisher, project_id, cp_topic_name)
-
-  logging.info("Publishing transactions to: %s", tx_path)
-  logging.info("Publishing coupons to: %s", cp_path)
-
-  households = ["1", "13", "42", "99", "125"]
-  tx_counter = 27601281000
-  published_count = 0
-
-  while True:
-    hh = random.choice(households)
-    tx_counter += 1
-    tx_items, cp_items = generate_synthetic_session_events(hh, tx_counter)
-
-    # Publish transactions
-    for tx in tx_items:
-      payload = json.dumps(tx).encode("utf-8")
-      future = publisher.publish(tx_path, payload)
-      logging.info("Published tx [hh=%s, tx=%s]: msg_id=%s", hh,
-                   tx["transaction_id"], future.result())
-
-    # Publish coupons
-    for cp in cp_items:
-      payload = json.dumps(cp).encode("utf-8")
-      future = publisher.publish(cp_path, payload)
-      logging.info("Published coupon [hh=%s, tx=%s]: msg_id=%s", hh,
-                   cp["transaction_id"], future.result())
-
-    # Error injection test (DLQ verification)
-    if inject_errors and random.random() < 0.2:
-      corrupt_payload = b"NOT_VALID_JSON_{broken: true"
-      future = publisher.publish(tx_path, corrupt_payload)
-      logging.info("Injected malformed transaction DLQ test payload: msg_id=%s",
-                   future.result())
-
-    published_count += len(tx_items) + len(cp_items)
-    if not continuous and published_count >= count:
-      break
-
-    await asyncio.sleep(interval)
-
+__all__ = [
+    "generate_synthetic_session_events",
+    "get_topic_path",
+    "publish_events_to_pubsub",
+]
 
 if __name__ == "__main__":
-  logging.basicConfig(
-      level=logging.INFO, format="%(asctime)s [%(levelname)s] %(message)s")
-  parser = argparse.ArgumentParser(
-      description="Publish Customer Data Platform sessions and events to Pub/Sub."
-  )
-  parser.add_argument(
-      "--project_id",
-      default=os.environ.get("PROJECT"),
-      help="GCP Project ID (defaults to $PROJECT)",
-  )
-  parser.add_argument(
-      "--transactions_topic",
-      default=os.environ.get("TRANSACTIONS_TOPIC"),
-      help="Transactions Pub/Sub topic name or path",
-  )
-  parser.add_argument(
-      "--coupons_topic",
-      default=os.environ.get("COUPON_REDEMPTION_TOPIC"),
-      help="Coupons Pub/Sub topic name or path",
-  )
-  parser.add_argument(
-      "--continuous",
-      action="store_true",
-      help="Continuously stream synthetic sessions until cancelled",
-  )
-  parser.add_argument(
-      "--interval",
-      type=float,
-      default=1.0,
-      help="Sleep interval in seconds between published customer sessions",
-  )
-  parser.add_argument(
-      "--count",
-      type=int,
-      default=20,
-      help="Total events to publish when not running continuously",
-  )
-  parser.add_argument(
-      "--inject_errors",
-      action="store_true",
-      help="Inject malformed payloads to verify Dead-Letter Queue (DLQ) processing",
-  )
-  args = parser.parse_args()
-
-  asyncio.run(
-      publish_events_to_pubsub(
-          project_id=args.project_id,
-          transactions_topic=args.transactions_topic,
-          coupons_topic=args.coupons_topic,
-          continuous=args.continuous,
-          interval=args.interval,
-          count=args.count,
-          inject_errors=args.inject_errors,
-      ))
+  from simulator.publisher import main  # pylint: disable=reimported
+  main()
