+++ pipelines/cdp/simulator/__init__.py
@@ -1,4 +1,4 @@
-#  Copyright 2025 Google LLC
+#  Copyright 2026 Google LLC
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -11,9 +11,4 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-apache-beam[gcp]==2.76.0  # Example, use your actual versions
-## Below dependencies are required if you run scripts/03_publish_events.py or simulator
-pandas
-fsspec
-gcsfs
+"""Customer Data Platform test data simulator package."""
